third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/html/semantics/forms/the-selectmenu-element/selectmenu-many-options.tentative-expected.txt

This is a testharness.js-based test.
FAIL The popover should be bottom left positioned assert_equals: expected 36 but got 20
Harness: the test ran to completion.
